requirements.txt
torch
numpy
nibabel
SimpleITK
pandas
matplotlib
tqdm
einops
pyyaml
MedPy
batchgenerators
ml-collections
tensorboardX
